Implement explicit scenario choice across package.
Introduce a scenario argument to calcCapital, calcFeDemand, calcFeDemandBuildings, calcFloorspace, calcTheil, and calcEmiTarget. The larger picture is the attempt to step away from a default set of scenarios. The choice of scenarios should be explicit. In the case of the input data pipeline, it is made in fullREMIND and then passed on to the different calcOutput functions. Hopefully this will help us control/drop/introduce new scenarios more easily in the future.
Move SSP2_NAV_all duplication to mrindustry.
Set NAs in CO2 prices from ExpertGuess to 0.
Drop unused sources argument in calcEmiTarget.
Drop unused scenario_proj argument in calcFE.
General refactoring.